src/outros/requirements.txt
# Edge Video/Audio Converter - Dependências

# Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# Processamento de Áudio/Vídeo
openai-whisper==20231117
mediapipe==0.10.8

# Utilitários
pathlib2==2.3.7
uuid==1.30

# Dependências adicionais do Whisper
torch==2.1.0
torchaudio==2.1.0
numpy==1.24.3
scipy==1.11.4

# Dependências do MediaPipe
opencv-python==4.8.1.78
protobuf==4.25.1

# Processamento de dados
Pillow==10.1.0

# Note: FFmpeg deve ser instalado separadamente
# Windows: winget install FFmpeg
# Linux: sudo apt install ffmpeg
# macOS: brew install ffmpeg
